Move "throw" helper methods to Builder class; fix ensureRegister()
