+++ javatools/src/main/java/org/xvm/asm/op/MoveCast.java
@@ -129,7 +129,7 @@ public class MoveCast
     public int build(BuildContext bctx, CodeBuilder code) {
         RegisterInfo regFrom = bctx.loadArgument(code, m_nFromValue);
         TypeConstant typeTo  = regFrom.type().combine(bctx.pool(), bctx.getTypeConstant(m_nToType));
-        bctx.generateCheckCast(code, typeTo);
+        bctx.builder.generateCheckCast(code, typeTo);
         bctx.storeValue(code, m_nToValue, typeTo);
         return -1;
     }
